Beta 01 - Corrections and functionalities
This build corrected some flaws that occurred in the previous build, such as crashing the application.

The Move Files button is now available and functional, it creates folders according to the files listed in the directory, moves the original files into the new folder if authorized, and when available, copies extra files within each folder.

This version can already be used and tested.