Log Generation
In this build the software creates a file called Log.txt in the original folder. This file contains the names of all files that are numbered in ascending order.

Still in this same file, in the last line is introduced a JSON with all the names already formatted and ready to be used.